Release v5.2.0: Star rating, quick notes, PROs/CONs tags, attachment notes
Features:

- Star rating control (5-star clickable, supports decimals)

- Quick notes chips (JSON-stored note chips with add/remove)

- PROs/CONs tag system (predefined + custom tags per property)

- Attachment notes column (multiline text for each attachment)

- Colorful UI icons for feature sections

Technical:

- PropertyTag & TagType models with many-to-many relationship

- ITagService & TagService for tag CRUD operations

- UpdateAttachmentAsync for saving attachment notes

- StarRatingControl custom WPF control

- Database schema updates with predefined tag seeding

Documentation:

- ARCHITECTURE.md comprehensive documentation

- Updated README.md with version history

+++ HomeBuyingApp.Infrastructure/Services/CsvService.cs
@@ -81,7 +81,7 @@ namespace HomeBuyingApp.Infrastructure.Services
                         PropertyTaxRate = ParseDecimal(values[14]),
                         Comments = values[15],
                         Notes = values[16],
-                        Rating = ParseInt(values[17]),
+                        Rating = ParseDecimal(values[17]),
                         LookAt = ParseBool(values[18]),
                         Interest = ParseBool(values[19]),
                         IsArchived = ParseBool(values[20])
